Seperating instructions into namespaces
While more intricate internally the resulting code will
be less error-prone. Since each instruction will supply
its own wrapper around the DecomposedRepresentation class
ArrayIndexOutOfBounds will be rarer.

Furthermore, the specification of an instruction became more clean and
represented in only one enum as opposed to spread out with the opcode
in one place, the pattern in another, the identifying characteristics
in a third.  It is all ine one place now.